+++ core-ng/src/main/java/core/framework/api/async/Executor.java
@@ -7,9 +7,5 @@ import java.util.concurrent.Future;
  * @author neo
  */
 public interface Executor {
-    <T> Batch<T> batch(String action);
-
-    <T> Batch<T> batch(String action, int maxConcurrentHandlers);
-
     <T> Future<T> submit(String action, Callable<T> task);
 }