+++ core-ng/src/main/java/core/framework/impl/async/ExecutorImpl.java
@@ -1,6 +1,5 @@
 package core.framework.impl.async;
 
-import core.framework.api.async.Batch;
 import core.framework.api.async.Executor;
 import core.framework.impl.log.ActionLog;
 import core.framework.impl.log.LogManager;
@@ -9,17 +8,14 @@ import org.slf4j.LoggerFactory;
 
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.atomic.AtomicInteger;
 
 /**
  * @author neo
  */
 public final class ExecutorImpl implements Executor {
     private final Logger logger = LoggerFactory.getLogger(ExecutorImpl.class);
-    private final ExecutorService executorService = Executors.newFixedThreadPool(Runtime.getRuntime().availableProcessors() * 4, new ExecutorThreadFactory());
+    private final ExecutorService executor = ThreadPools.cachedThreadPool(Runtime.getRuntime().availableProcessors() * 2, "executor-");
     private final LogManager logManager;
 
     public ExecutorImpl(LogManager logManager) {
@@ -28,35 +24,31 @@ public final class ExecutorImpl implements Executor {
 
     public void stop() {
         logger.info("stop executor");
-        executorService.shutdown();
-    }
-
-    @Override
-    public <T> Batch<T> batch(String action) {
-        return batch(action, Runtime.getRuntime().availableProcessors() * 4);
-    }
-
-    @Override
-    public <T> Batch<T> batch(String action, int maxConcurrentHandlers) {
-        return new Batch<>(action, maxConcurrentHandlers, this);
+        executor.shutdown();
     }
 
     @Override
     public <T> Future<T> submit(String action, Callable<T> task) {
-        // do not log intentionally, so in batch pattern where one batch submit task for each item, the batch process won't reach max trace log line limit.
         ActionLog parentActionLog = logManager.currentActionLog();
-        String childAction = parentActionLog != null ? parentActionLog.action + "/" + action : action;
-        String refId = parentActionLog != null ? parentActionLog.refId() : null;
-        boolean trace = parentActionLog != null && parentActionLog.trace;
-        return executorService.submit(new Task<>(task, logManager, childAction, refId, trace));
-    }
-
-    private static class ExecutorThreadFactory implements ThreadFactory {
-        private final AtomicInteger count = new AtomicInteger(1);
-
-        @Override
-        public Thread newThread(Runnable runnable) {
-            return new Thread(runnable, "executor-thread-" + count.getAndIncrement());
-        }
+        String taskAction = parentActionLog.action + "/" + action;
+        String refId = parentActionLog.refId();
+        boolean trace = parentActionLog.trace;
+        return executor.submit(() -> {
+            try {
+                logManager.begin("=== task execution begin ===");
+                ActionLog actionLog = logManager.currentActionLog();
+                actionLog.refId(refId);
+                actionLog.action(taskAction);
+                if (trace) {
+                    actionLog.trace = true;
+                }
+                return task.call();
+            } catch (Throwable e) {
+                logManager.logError(e);
+                throw e;
+            } finally {
+                logManager.end("=== task execution end ===");
+            }
+        });
     }
 }