+++ core-ng/src/main/java/core/framework/impl/queue/RabbitMQListener.java
@@ -2,7 +2,6 @@ package core.framework.impl.queue;
 
 import com.rabbitmq.client.AMQP;
 import com.rabbitmq.client.QueueingConsumer;
-import core.framework.api.async.Executor;
 import core.framework.api.module.MessageHandlerConfig;
 import core.framework.api.queue.Message;
 import core.framework.api.queue.MessageHandler;
@@ -10,6 +9,7 @@ import core.framework.api.util.Exceptions;
 import core.framework.api.util.Maps;
 import core.framework.api.util.Strings;
 import core.framework.api.util.Threads;
+import core.framework.impl.async.ThreadPools;
 import core.framework.impl.json.JSONReader;
 import core.framework.impl.log.ActionLog;
 import core.framework.impl.log.LogManager;
@@ -19,13 +19,13 @@ import org.slf4j.LoggerFactory;
 
 import java.time.Duration;
 import java.util.Map;
-import java.util.concurrent.Semaphore;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
  * @author neo
  */
-public class RabbitMQListener implements MessageHandlerConfig {
+public final class RabbitMQListener implements MessageHandlerConfig {
     static final String HEADER_TRACE = "trace";
     static final String HEADER_CLIENT_IP = "clientIP";
 
@@ -34,16 +34,14 @@ public class RabbitMQListener implements MessageHandlerConfig {
     private final AtomicBoolean stop = new AtomicBoolean(false);
     private final Thread listenerThread;
     private final String queue;
-    private final Executor executor;
     private final LogManager logManager;
     private final MessageValidator validator;
     private final Map<String, MessageHandler> handlers = Maps.newHashMap();
     private final Map<String, JSONReader> readers = Maps.newHashMap();
-    private int maxConcurrentHandlers = 10;
-    private Semaphore semaphore;
+    private int poolSize = Runtime.getRuntime().availableProcessors() * 2;
+    private ExecutorService handlerThreadPool;
 
-    public RabbitMQListener(RabbitMQ rabbitMQ, String queue, Executor executor, MessageValidator validator, LogManager logManager) {
-        this.executor = executor;
+    public RabbitMQListener(RabbitMQ rabbitMQ, String queue, MessageValidator validator, LogManager logManager) {
         this.queue = queue;
         this.validator = validator;
         this.logManager = logManager;
@@ -51,8 +49,11 @@ public class RabbitMQListener implements MessageHandlerConfig {
         listenerThread = new Thread(() -> {
             logger.info("rabbitMQ listener started, queue={}", queue);
             while (!stop.get()) {
-                try (RabbitMQConsumer consumer = rabbitMQ.consumer(queue, maxConcurrentHandlers * 2)) { // prefetch one more for each handler to improve throughput
-                    pullMessages(consumer);
+                try (RabbitMQConsumer consumer = rabbitMQ.consumer(queue, poolSize * 2)) { // prefetch one more for each handler to improve throughput
+                    while (!stop.get()) {
+                        QueueingConsumer.Delivery delivery = consumer.nextDelivery();
+                        handlerThreadPool.submit(() -> handle(consumer, delivery));
+                    }
                 } catch (Throwable e) {
                     if (!stop.get()) {  // if not initiated by shutdown, exception types can be ShutdownSignalException, InterruptedException
                         logger.error("failed to pull message, retry in 30 seconds", e);
@@ -60,8 +61,7 @@ public class RabbitMQListener implements MessageHandlerConfig {
                     }
                 }
             }
-        });
-        listenerThread.setName("rabbitMQ-listener-" + queue);
+        }, "rabbitMQ-" + queue + "-listener");
     }
 
     @Override
@@ -77,37 +77,13 @@ public class RabbitMQListener implements MessageHandlerConfig {
     }
 
     @Override
-    public MessageHandlerConfig maxConcurrentHandlers(int maxConcurrentHandlers) {
-        this.maxConcurrentHandlers = maxConcurrentHandlers;
+    public MessageHandlerConfig poolSize(int poolSize) {
+        this.poolSize = poolSize;
         return this;
     }
 
-    private void pullMessages(RabbitMQConsumer consumer) throws InterruptedException {
-        String action = "queue/" + queue;
-        while (!stop.get()) {
-            QueueingConsumer.Delivery delivery = consumer.nextDelivery();
-            semaphore.acquire(); // acquire permit right before submit, to avoid permit failing to release back due to exception in between
-            executor.submit(action, () -> {
-                try {
-                    process(delivery);
-                    return null;
-                } finally {
-                    complete(consumer, delivery);
-                }
-            });
-        }
-    }
-
-    private void complete(RabbitMQConsumer consumer, QueueingConsumer.Delivery delivery) {
-        try {
-            consumer.acknowledge(delivery.getEnvelope().getDeliveryTag());
-        } finally {
-            semaphore.release(); // make sure release semaphore at end, not release before acknowledge is to reduce the num of thread to be created in pool
-        }
-    }
-
     public void start() {
-        semaphore = new Semaphore(maxConcurrentHandlers, false);
+        handlerThreadPool = ThreadPools.cachedThreadPool(poolSize, "rabbitMQ-" + queue + "-handler-");
         listenerThread.start();
     }
 
@@ -115,27 +91,50 @@ public class RabbitMQListener implements MessageHandlerConfig {
         logger.info("stop rabbitMQ listener, queue={}", queue);
         stop.set(true);
         listenerThread.interrupt();
+        handlerThreadPool.shutdown();
     }
 
-    private <T> void process(QueueingConsumer.Delivery delivery) throws Exception {
+    private Void handle(RabbitMQConsumer consumer, QueueingConsumer.Delivery delivery) throws Exception {
+        try {
+            logManager.begin("=== message handling begin ===");
+            handle(delivery);
+            return null;
+        } catch (Throwable e) {
+            logManager.logError(e);
+            throw e;
+        } finally {
+            acknowledge(consumer, delivery);
+            logManager.end("=== message handling end ===");
+        }
+    }
+
+    private void acknowledge(RabbitMQConsumer consumer, QueueingConsumer.Delivery delivery) {
+        try {
+            consumer.acknowledge(delivery.getEnvelope().getDeliveryTag());
+        } catch (Throwable e) {
+            logger.error(e.getMessage(), e);
+        }
+    }
+
+    private <T> void handle(QueueingConsumer.Delivery delivery) throws Exception {
         ActionLog actionLog = logManager.currentActionLog();
+        actionLog.action("queue/" + queue);
 
         AMQP.BasicProperties properties = delivery.getProperties();
-        byte[] body = delivery.getBody();
-
         String messageType = properties.getType();
         actionLog.context("messageType", messageType);
 
+        byte[] body = delivery.getBody();
         logger.debug("body={}", LogParam.of(body));
 
-        if (Strings.isEmpty(messageType)) throw new Error("messageType must not be empty");
+        if (Strings.isEmpty(messageType)) throw new Error("message type must not be empty");
 
         actionLog.refId(properties.getCorrelationId());
 
         Map<String, Object> headers = properties.getHeaders();
         if (headers != null) {
             if ("true".equals(String.valueOf(headers.get(HEADER_TRACE)))) {
-                logManager.triggerTraceLog();
+                actionLog.trace = true;
             }
 
             Object clientIP = headers.get(HEADER_CLIENT_IP);
@@ -151,15 +150,14 @@ public class RabbitMQListener implements MessageHandlerConfig {
 
         @SuppressWarnings("unchecked")
         JSONReader<T> reader = readers.get(messageType);
-        if (reader == null) {
-            throw Exceptions.error("unknown message type, messageType={}", messageType);
-        }
+        if (reader == null) throw Exceptions.error("unknown message type, messageType={}", messageType);
         T message = reader.fromJSON(body);
         validator.validate(message);
 
         @SuppressWarnings("unchecked")
         MessageHandler<T> handler = handlers.get(messageType);
         actionLog.context("handler", handler.getClass().getCanonicalName());
+
         handler.handle(message);
     }
 }