+++ core-ng/src/main/java/core/framework/impl/web/HTTPServerHandler.java
@@ -72,7 +72,7 @@ public class HTTPServerHandler implements HttpHandler {
 
             // trigger trace after action is determined due to trace log use action as part of path, is there better way?
             if ("true".equals(headers.getFirst(HEADER_TRACE))) {
-                logManager.triggerTraceLog();
+                actionLog.trace = true;
             }
 
             webContext.initialize(request);